projects/03.5-outreach-assistant/requirements.txt
python-dotenv==1.1.1
google-api-python-client>=2.170.0,<3
google-auth>=2.35.0,<3
google-auth-oauthlib>=1.2.0,<2
pytest>=8.0.0,<9
